+++ changelog.md
@@ -26,3 +26,4 @@
 - Signature of ImageInterface::crop() changed `offset_x` is no `x` and `offset_y` is now `y`
 - Signature of ImageInterface::place() changed `offset_x` is no `x` and `offset_y` is now `y`
 - EncodedImageInterface::toDataUri() now returns `DataUriInterface` instead of `string´
+- ProfileInterface requires implementation of `::fromPath()`